requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.1
attrs==25.4.0
Automat==25.4.16
beautifulsoup4==4.14.2
billiard==4.2.2
brotli==1.2.0
bytecode==0.17.0
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.3
cssselect==1.3.0
cssselect2==0.8.0
cycler==0.12.1
datadog==0.52.1
ddtrace==3.18.1
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.2.8
django-extensions==4.1
docstring_parser==0.17.0
dotenv==0.9.9
envier==0.6.1
executing==2.2.1
filelock==3.20.0
flower==2.0.1
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httpx==0.28.1
humanize==4.14.0
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.0
incremental==24.7.2
ipython==9.7.0
ipython_pygments_lexers==1.1.1
itemadapter==0.12.2
itemloaders==1.3.2
jedi==0.19.2
jiter==0.12.0
jmespath==1.0.1
JSON-log-formatter==1.1.1
kiwisolver==1.4.9
kombu==5.5.4
langdetect==1.0.9
legacy-cgi==2.6.4
lxml==6.0.2
lxml_html_clean==0.4.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
numpy==2.3.5
opentelemetry-api==1.38.0
packaging==25.0
parsel==1.10.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
playwright==1.56.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
Protego==0.5.0
protobuf==6.33.1
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyDispatcher==2.0.7
pydyf==0.11.0
pyee==13.0.0
Pygments==2.19.2
pyOpenSSL==25.3.0
pyparsing==3.2.5
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
queuelib==1.8.0
readability-lxml==0.8.4.1
redis==7.0.1
requests==2.32.5
requests-file==3.0.1
Scrapy==2.13.3
scrapy-playwright==0.0.44
scrapy-redis==0.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
text-unidecode==1.3
textstat==0.7.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tldextract==5.3.0
tornado==6.5.2
traitlets==5.14.3
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
w3lib==2.3.1
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
wrapt==2.0.1
zipp==3.23.0
zope.interface==8.1.1
zopfli==0.4.0
